+++ docs/Changelog.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 * enh: config file union support
+* fix: invoking gui interface twice
 
 ## 1.2.1 (2025-10-24)
 * deps (tyro): ready for 0.10